io refactoring
